Enable the test for the basic evaluator.
