+++ Evaluators/BasicEvaluator.Macros.cs
@@ -28,21 +28,6 @@ namespace MetaphysicsIndustries.Solus.Evaluators
 {
     public partial class BasicEvaluator
     {
-        public Expression CallMacro(DeriveOperator mm, IMathObject[] args,
-            SolusEnvironment env)
-        {
-            return CallFunction(mm, args, env);
-        }
-
-        public Expression CallFunction(DeriveOperator f, IMathObject[] args,
-            SolusEnvironment env)
-        {
-            var derive = new DerivativeTransformer();
-            var expr = (Expression)args[0];
-            var v = ((VariableAccess)args[1]).VariableName;
-            return derive.Transform(expr, new VariableTransformArgs(v));
-        }
-
         public Expression CallMacro(IfMacro mm, Expression[] args,
             SolusEnvironment env)
         {
